Contact section
Changed min-height for the sections, so that the first 2 have a min-height: 80vh and the third a min-height: 90vh (plus border values), more apropriate since now that is 100% of their screen space.

Made the contact section. Now I need to add functionality to the form and add the geologist shining a lantern, which is a design touch.