Fixing bug with responsive design
Simple fix with @media and removing the fixed height of the #social div at the end of the page. Now, the geologist is gone on screens of width lower than 768px, and the #social can adjust its height based on the flex-flow, thus removing bugs that were happening in the sticky layer sections in front.